requirements.txt
aiohttp==3.7.3
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
beautifulsoup4==4.9.3
braceexpand==0.1.6
bs4==0.0.1
buttons==0.1.9
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
DateTime==4.3
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@6f2b873bf0d28903eb752aa1166b3aac26dc9007
disputils==0.1.3
enhanced-dpy==1.6.0.6
futures==3.1.1
goslate==1.5.1
humanize==3.2.0
idna==2.10
import-expression==1.1.4
jishaku==2.0.0
motor==2.3.0
multidict==5.1.0
mystbin.py==1.0.1
psutil==5.8.0
pymongo==3.11.2
pytz==2020.5
ratelimiter==1.2.0.post0
requests==2.25.1
six==1.15.0
soupsieve==2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
zope.interface==5.2.0